requirements.txt
duckduckgo-search==8.0.2
huggingface-hub[hf_xet]==0.30.2
llama-cpp-python==0.3.8
llama-cpp-agent==0.2.35
llama-index==0.12.34
regex==2024.11.6
streamlit==1.45.0
wikitextparser==0.56.4